Changes to add tenant to urls
Bouncer v2 requires that the tenant be provided in the URL.  These
modifications allow the tenant to be specified on the client, and
changes how the URL is built. `buildPath` was removed from utils.js (and
utils.js was removed) and relocated into client.js to make use of the
new tenant property.  `buildUrl` and `buildPath` can probably now be
combined and/or simplified. GET/POST/etc... methods now take an optional
override tenant ID, this was used to simplify/unify the identity URL
paths.