Changes from code review
These were the changes from our JS client code review.

* Ensure path variable segments can't be empty
* Change SSCProxy to ServiceClient
* Change *Proxy to *Service
* Change Splunk to SplunkSSC (interim name)

In order to ensure that the path variable segments aren't empty I
changed from a string path fragment to an array of path elements- that
way we can validate that none are empty.  I _think_ that's a decent
approach, but if anyone hates it we can revisit.