add getResults, resultsObservable, statusObservable
Added an observable that will periodically poll a job and return it's
status.  Also added a method to fetch results, either by a window or by
retrieving all results. Also added a resultObserver method- this works
differently than the previous eventObserver in that it does *not* wait
for the job to complete, rather it will start sending a full set of
results every poll interval (or a window), allowing for our 'build up'
behavior as seen in enterprise Splunk