Various small bugfixes (#93)
* Fixed a bug with passing attrs to eventObserver

* Fixed a bug in the new search docs.

* Didn't actually specify pollInterval in setTimeout.

* Renamed search args to use camelCase

To match amendment to search spec

* Reject waitForJob when dispatchState is 'FAILED'

* add getResults, resultsObservable, statusObservable

Added an observable that will periodically poll a job and return it's
status.  Also added a method to fetch results, either by a window or by
retrieving all results. Also added a resultObserver method- this works
differently than the previous eventObserver in that it does *not* wait
for the job to complete, rather it will start sending a full set of
results every poll interval (or a window), allowing for our 'build up'
behavior as seen in enterprise Splunk

* Removed co lib in favor of async/await

* Fixed error when cancelling search

Was allowing an error to escape a promise, makeing it uncatchable.
Also changed the error to explicitly mention cancellation.

* s/NewSearchConfig/PostJobsRequest/

* HEC Raw objects were getting double-stringified

Leading to values with extra quote marks

* Missed a 'co' import

* Moved utility function out of public class

Moved some babel config around to make babel happy.